[add] Add module loader_lib that loads App and Lib at the same time.
[examples/loader_lib] has been added. This application can be used to
load such a bin file: a simple package of the application (App) and the
mock dynamic library (Lib) that the application depends on. [loader_lib]
will load App and Lib at the same time, and modify PLT and GOT so that
the symbols used in Lib in App are relocated to the location of Lib in
memory. loader_lib will start execution from a specific function in Lib,
and after completing specific settings, jump to the entrance of the App
for execution. The parts of Lib that depend on ArceOS will be jumped
through the ABI Table. Lib currently supports some common functions.